+++ src/dataPowers/Life.java
@@ -15,7 +15,7 @@ public class Life extends PowerUp
 	// determina se l'oggetto ha raggiunto terra
 	private boolean arrived = false;
 	
-	public Life( int x, int y, int ray, double maxH ) throws SlickException
+	public Life( int x, int y, float ray, double maxH ) throws SlickException
 		{
 			super( "life" );
